+++ src/web/nextui/src/app/setup/page.css
